server/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.14.0
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asteroid-filterbanks==0.4.0
attrs==24.2.0
audioread==3.0.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
coards==1.0.5
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
decorator==4.4.2
distro==1.9.0
docopt==0.6.2
einops==0.8.0
Extract==1.0
fastapi==0.115.4
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.0.3
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.27.0
HyperPyYAML==1.2.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
kiwisolver==1.4.8
langchain==0.3.7
langchain-core==0.3.15
langchain-openai==0.2.6
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.140
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.5.0.post0
lightning-utilities==0.11.9
llvmlite==0.43.0
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
narwhals==1.13.2
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
openai==1.54.3
openai-whisper==20240930
optuna==4.1.0
orjson==3.10.11
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pinecone==5.3.1
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3
proglog==0.1.10
propcache==0.2.0
protobuf==5.28.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydap==3.5.1
pydeck==0.9.1
pyee==12.0.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-ffmpeg==2.0.12
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.21.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.0
scipy==1.14.1
semver==3.0.2
sentencepiece==0.2.0
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.36
starlette==0.41.2
streamlit==1.40.0
Subclip==0.0.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.8.0
toml==0.10.2
torch==2.5.1
torch-audiomentations==0.11.1
torch_pitch_shift==1.2.5
torchaudio==2.5.1
torchmetrics==1.6.1
tornado==6.4.1
tqdm==4.66.6
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.1
yarl==1.17.1
